Upgrade to Boot 2.0.x
Based on Janne's PR https://github.com/spring-cloud/spring-cloud-dataflow/pull/2523

+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/controller/AppRegistryController.java
@@ -134,7 +134,7 @@ public class AppRegistryController implements ResourceLoaderAware {
 			// building new page.
 			// all this is done because we don't use a proper repository which would
 			// handle all these automatically.
-			int offset = 0;
+			long offset = 0;
 			int page = 0;
 			if (pageable.getOffset() <= to) {
 				offset = pageable.getOffset();
@@ -143,7 +143,7 @@ public class AppRegistryController implements ResourceLoaderAware {
 			else if (pageable.getOffset() + pageable.getPageSize() <= to) {
 				offset = pageable.getOffset();
 			}
-			pagedRegistrations = new PageImpl<>(appRegistrations.subList(offset, (int) to),
+			pagedRegistrations = new PageImpl<>(appRegistrations.subList((int) offset, (int) to),
 					new PageRequest(page, pageable.getPageSize()), appRegistrations.size());
 		}
