+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/repository/support/SqlServerPagingQueryProvider.java
@@ -28,7 +28,7 @@ public class SqlServerPagingQueryProvider extends AbstractSqlPagingQueryProvider
 
 	@Override
 	public String getPageQuery(Pageable pageable) {
-		int offset = pageable.getOffset() + 1;
+		long offset = pageable.getOffset() + 1;
 		return generateRowNumSqlQueryWithNesting(getSelectClause(), false,
 				"TMP_ROW_NUM >= " + offset + " AND TMP_ROW_NUM < " + (offset + pageable.getPageSize()));
 	}
